ReadingProgress: move all logic to HOC
